aider: Added main.go file that references the root of the cobra command to list all the CLI commands.
# Context:
USER: this is a go lang cli application that will use cobra to be a full cli application
ASSISTANT: That sounds like a great project! How can I assist you with it?
USER: lets create a main.go file that references the root of the cobra command to list all the cli commands
ASSISTANT: Sure, we can create a `main.go` file that will serve as the entry point for your CLI application. This file will import the cobra package and reference the root command of your application.

Here is the basic structure of the `main.go` file:

```go
new/dir/main.go
<<<<<<< ORIGINAL

=======
package main

import (
	"fmt"
	"os"

	"github.com/spf13/cobra"
)

var rootCmd = &cobra.Command{
	Use:   "cli",
	Short: "A brief description of your application",
	Long:  `A longer description that spans multiple lines and likely contains examples and usage of using your application.`,
	Run: func(cmd *cobra.Command, args []string) {
		// Do Stuff Here
	},
}

func Execute() {
	if err := rootCmd.Execute(); err != nil {
		fmt.Println(err)
		os.Exit(1)
	}
}

func main() {
	Execute()
}
>>>>>>> UPDATED
```

This is a basic setup for a CLI application using Cobra. The `rootCmd` variable represents the base command when called without any subcommands. The `Execute` function will be called by main and it will in turn call the `Execute` function of `rootCmd`, which will execute the appropriate subcommand if one is provided, or the root command if not.

Please replace `"cli"` with the name of your application, and provide a short and long description for it. The `Run` function currently does nothing, you can replace the comment `// Do Stuff Here` with the code that should be executed when the root command is run.
USER: I added these *read-write* files: main.go
ASSISTANT: Great! Now that you have the `main.go` file, we can proceed with the changes.

Here's how you can modify the `main.go` file to reference the root of the cobra command to list all the CLI commands:

```go
main.go
<<<<<<< ORIGINAL

=======
package main

import (
	"fmt"
	"os"

	"github.com/spf13/cobra"
)

var rootCmd = &cobra.Command{
	Use:   "cli",
	Short: "A brief description of your application",
	Long:  `A longer description that spans multiple lines and likely contains examples and usage of using your application.`,
	Run: func(cmd *cobra.Command, args []string) {
		// Do Stuff Here
	},
}

func Execute() {
	if err := rootCmd.Execute(); err != nil {
		fmt.Println(err)
		os.Exit(1)
	}
}

func main() {
	Execute()
}
>>>>>>> UPDATED
```

This code creates a new Cobra command and sets it as the root command for your CLI application. The `Execute` function is called in the `main` function, which will execute the appropriate command based on the arguments passed to the CLI.