aider: Added a new command called `kubernetes` to the `rootCmd` in the `main.go` file, with multiple flags and corresponding environment variables.
# Context:
USER: now we need to make a new command called kubernetes. The command will have multiple input variables which are all strings like akeyless-source-token and akeyless-destination-token and source-gateway-url and destination-gateway-url and filter-config-file-path and each of these should be able to be set as an environment variable like AKEYLESS_SOURCE_TOKEN for the akeyless-source-token flag. the description for this command will be to migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster.
ASSISTANT: To implement this new command, we need to modify the `main.go` file. However, it's currently read-only. Could you please make it read-write? The full path is likely `./main.go`.
USER: I added these *read-write* files: main.go
ASSISTANT: Great! Now we can proceed with the changes.

We will add a new command called `kubernetes` to the `rootCmd` in the `main.go` file. This command will have multiple flags, each corresponding to an input variable. The flags will be `akeyless-source-token`, `akeyless-destination-token`, `source-gateway-url`, `destination-gateway-url`, and `filter-config-file-path`. Each flag will also have a corresponding environment variable.

Here are the changes:

```go
main.go
<<<<<<< ORIGINAL
var rootCmd = &cobra.Command{
	Use:   "agmigrator",
	Short: "Akeyless Gateway Configuration Migrator",
	Long:  `agmigrator is a CLI tool for migrating configuration from one Akeyless Gateway cluster to another.`,
	Run: func(cmd *cobra.Command, args []string) {
		// Do Stuff Here
	},
}
=======
var (
	akeylessSourceToken      string
	akeylessDestinationToken string
	sourceGatewayURL         string
	destinationGatewayURL    string
	filterConfigFilePath     string
)

var rootCmd = &cobra.Command{
	Use:   "agmigrator",
	Short: "Akeyless Gateway Configuration Migrator",
	Long:  `agmigrator is a CLI tool for migrating configuration from one Akeyless Gateway cluster to another.`,
}

var kubernetesCmd = &cobra.Command{
	Use:   "kubernetes",
	Short: "Migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster",
	Run: func(cmd *cobra.Command, args []string) {
		// Do Stuff Here
	},
}

func init() {
	rootCmd.AddCommand(kubernetesCmd)

	kubernetesCmd.Flags().StringVar(&akeylessSourceToken, "akeyless-source-token", "", "Akeyless source token")
	kubernetesCmd.Flags().StringVar(&akeylessDestinationToken, "akeyless-destination-token", "", "Akeyless destination token")
	kubernetesCmd.Flags().StringVar(&sourceGatewayURL, "source-gateway-url", "", "Source gateway URL")
	kubernetesCmd.Flags().StringVar(&destinationGatewayURL, "destination-gateway-url", "", "Destination gateway URL")
	kubernetesCmd.Flags().StringVar(&filterConfigFilePath, "filter-config-file-path", "", "Filter config file path")

	viper.BindPFlag("akeyless-source-token", kubernetesCmd.Flags().Lookup("akeyless-source-token"))
	viper.BindPFlag("akeyless-destination-token", kubernetesCmd.Flags().Lookup("akeyless-destination-token"))
	viper.BindPFlag("source-gateway-url", kubernetesCmd.Flags().Lookup("source-gateway-url"))
	viper.BindPFlag("destination-gateway-url", kubernetesCmd.Flags().Lookup("destination-gateway-url"))
	viper.BindPFlag("filter-config-file-path", kubernetesCmd.Flags().Lookup("filter-config-file-path"))

	viper.AutomaticEnv()
}
>>>>>>> UPDATED
```

In the updated code, we first declare the variables that will hold the values of the flags. Then we define the `kubernetes` command and add it to the `rootCmd`. In the `init` function, we add the flags to the `kubernetes` command and bind them to the corresponding variables. We also bind the flags to their corresponding environment variables using `viper.BindPFlag` and `viper.AutomaticEnv`.