aider: Updated kubernetes command to get values from environment variables if flags are not set.
# Context:
USER: viper is not valid as we are using cobra
ASSISTANT: I apologize for the confusion. You're right, `viper` is not used with `cobra` in your current setup.

We can use `os.Getenv` to get the environment variables in the `Run` function of the `kubernetes` command. If the flag is not set, we will use the value from the environment variable.

Here are the changes:

```go
main.go
<<<<<<< ORIGINAL
var kubernetesCmd = &cobra.Command{
	Use:   "kubernetes",
	Short: "Migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster",
	Run: func(cmd *cobra.Command, args []string) {
		// Do Stuff Here
	},
}
=======
var kubernetesCmd = &cobra.Command{
	Use:   "kubernetes",
	Short: "Migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster",
	Run: func(cmd *cobra.Command, args []string) {
		if akeylessSourceToken == "" {
			akeylessSourceToken = os.Getenv("AKEYLESS_SOURCE_TOKEN")
		}
		if akeylessDestinationToken == "" {
			akeylessDestinationToken = os.Getenv("AKEYLESS_DESTINATION_TOKEN")
		}
		if sourceGatewayURL == "" {
			sourceGatewayURL = os.Getenv("SOURCE_GATEWAY_URL")
		}
		if destinationGatewayURL == "" {
			destinationGatewayURL = os.Getenv("DESTINATION_GATEWAY_URL")
		}
		if filterConfigFilePath == "" {
			filterConfigFilePath = os.Getenv("FILTER_CONFIG_FILE_PATH")
		}
		// Do Stuff Here
	},
}
>>>>>>> UPDATED
```

In the updated code, we check if each flag is set. If not, we get the value from the corresponding environment variable.