aider: Prepend "AKEYLESS_" to environment variables that do not start with "AKEYLESS".
# Context:
USER: that is great, now if the environment variable already starts with AKEYLESS then leave it be but if it does not start with AKEYLESS then please prepend the word AKEYLESS with an underscore to make it easier to see.
ASSISTANT: To implement this, we need to modify the `Run` function in the `kubernetes` command in the `main.go` file. We will check if the environment variable name starts with "AKEYLESS". If it doesn't, we will prepend "AKEYLESS_" to the name.

Here are the changes:

```go
main.go
<<<<<<< ORIGINAL
var kubernetesCmd = &cobra.Command{
	Use:   "kubernetes",
	Short: "Migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster",
	Run: func(cmd *cobra.Command, args []string) {
		if akeylessSourceToken == "" {
			akeylessSourceToken = os.Getenv("AKEYLESS_SOURCE_TOKEN")
		}
		if akeylessDestinationToken == "" {
			akeylessDestinationToken = os.Getenv("AKEYLESS_DESTINATION_TOKEN")
		}
		if sourceGatewayURL == "" {
			sourceGatewayURL = os.Getenv("SOURCE_GATEWAY_URL")
		}
		if destinationGatewayURL == "" {
			destinationGatewayURL = os.Getenv("DESTINATION_GATEWAY_URL")
		}
		if filterConfigFilePath == "" {
			filterConfigFilePath = os.Getenv("FILTER_CONFIG_FILE_PATH")
		}
		// Do Stuff Here
	},
}
=======
var kubernetesCmd = &cobra.Command{
	Use:   "kubernetes",
	Short: "Migrate kubernetes auth configs from a source Akeyless gateway cluster to a destination Akeyless gateway cluster",
	Run: func(cmd *cobra.Command, args []string) {
		if akeylessSourceToken == "" {
			akeylessSourceToken = getEnvVar("AKEYLESS_SOURCE_TOKEN")
		}
		if akeylessDestinationToken == "" {
			akeylessDestinationToken = getEnvVar("AKEYLESS_DESTINATION_TOKEN")
		}
		if sourceGatewayURL == "" {
			sourceGatewayURL = getEnvVar("SOURCE_GATEWAY_URL")
		}
		if destinationGatewayURL == "" {
			destinationGatewayURL = getEnvVar("DESTINATION_GATEWAY_URL")
		}
		if filterConfigFilePath == "" {
			filterConfigFilePath = getEnvVar("FILTER_CONFIG_FILE_PATH")
		}
		// Do Stuff Here
	},
}

func getEnvVar(name string) string {
	if !strings.HasPrefix(name, "AKEYLESS") {
		name = "AKEYLESS_" + name
	}
	return os.Getenv(name)
}
>>>>>>> UPDATED
```

In the updated code, we replace the direct calls to `os.Getenv` with calls to a new function `getEnvVar`. This function checks if the name starts with "AKEYLESS". If it doesn't, it prepends "AKEYLESS_" to the name before getting the environment variable.