aider: Added condition to check if the token is valid and return an error message if it is not.
# Context:
USER: inside the runValidateToken func is the return object of the validate token. if the IsValid property is false then the time difference will fail and so should be avoided with a message about the token NOT being valid.
ASSISTANT: To implement this change, we need to modify the `runValidateToken` function in the `main.go` file. We will add a condition to check if the `IsValid` property of the `validateToken` object is `false`. If it is, we will print a message about the token not being valid and return an error.

Here is the proposed change:

```go
main.go
<<<<<<< ORIGINAL
func runValidateToken(akeylessToken string, gatewayConfigURL string) error {
	akeylessService := factories.BuildAkeylessService(sourceGatewayConfigURL)
	validateToken, err := akeylessService.ValidateToken(context.Background(), akeylessSourceToken)
	if err != nil {
		fmt.Println("Unable to validate token at URL:", gatewayConfigURL, err)
		return err
	} else {
		// print line to indicate if token is valid and the token expiration time
		expirationTime, err := time.Parse("2006-01-02 15:04:05 -0700 MST", *validateToken.Expiration)
		if err != nil {
			fmt.Println("Unable to parse expiration time:", err)
			return err
		}
		// print the difference between the current time and the token expiration time, rounded to the nearest second
		fmt.Println("Token expires in:", time.Until(expirationTime).Round(time.Second))
	}
	return err
}
=======
func runValidateToken(akeylessToken string, gatewayConfigURL string) error {
	akeylessService := factories.BuildAkeylessService(sourceGatewayConfigURL)
	validateToken, err := akeylessService.ValidateToken(context.Background(), akeylessSourceToken)
	if err != nil {
		fmt.Println("Unable to validate token at URL:", gatewayConfigURL, err)
		return err
	} else {
		if !validateToken.IsValid {
			fmt.Println("Token is not valid")
			return errors.New("Token is not valid")
		}
		// print line to indicate if token is valid and the token expiration time
		expirationTime, err := time.Parse("2006-01-02 15:04:05 -0700 MST", *validateToken.Expiration)
		if err != nil {
			fmt.Println("Unable to parse expiration time:", err)
			return err
		}
		// print the difference between the current time and the token expiration time, rounded to the nearest second
		fmt.Println("Token expires in:", time.Until(expirationTime).Round(time.Second))
	}
	return err
}
>>>>>>> UPDATED
```

Please note that you need to import the `errors` package to use `errors.New`. If it's not already imported, add `import "errors"` at the top of your file.